+++ packages/ng-table-builder/src/components/ngx-header/ngx-header.component.ts
@@ -1,4 +1,7 @@
-import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output, ViewEncapsulation } from '@angular/core';
+import { coerceBoolean } from '@angular-ru/common/coercion';
+import { AttributeBoolean } from '@angular-ru/common/decorators';
+import { InputBoolean } from '@angular-ru/common/typings';
 
 import { TableContentDirective } from '../../directives/table-content.directive';
 
@@ -8,4 +11,14 @@ import { TableContentDirective } from '../../directives/table-content.directive'
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None
 })
-export class NgxHeaderComponent extends TableContentDirective {}
+export class NgxHeaderComponent extends TableContentDirective {
+    @Output() public readonly expandedChange: EventEmitter<boolean> = new EventEmitter<boolean>();
+    @AttributeBoolean() @Input() public expanded: InputBoolean = true;
+    @AttributeBoolean() @Input() public expandablePanel: InputBoolean = false;
+    @AttributeBoolean() @Input() public hideToggle: InputBoolean = false;
+
+    public toggleExpand(): void {
+        this.expanded = !coerceBoolean(this.expanded);
+        this.expandedChange.emit(this.expanded);
+    }
+}