+++ packages/ng-table-builder/src/table-builder.component.ts
@@ -20,6 +20,8 @@ import {
     ViewEncapsulation
 } from '@angular/core';
 import { fadeInLinearAnimation } from '@angular-ru/common/animations';
+import { hasItems } from '@angular-ru/common/array';
+import { coerceBoolean } from '@angular-ru/common/coercion';
 import { Any, DeepPartial, Nullable, PlainObjectOf, SortOrderType } from '@angular-ru/common/typings';
 import { checkValueIsFilled, detectChanges, isFalse, isFalsy, isNil, isNotNil } from '@angular-ru/common/utils';
 import { EMPTY, fromEvent, Observable, Subject } from 'rxjs';
@@ -28,8 +30,10 @@ import { catchError, takeUntil } from 'rxjs/operators';
 import { AbstractTableBuilderApiDirective } from './abstract-table-builder-api.directive';
 import { NgxColumnComponent } from './components/ngx-column/ngx-column.component';
 import { TABLE_GLOBAL_OPTIONS } from './config/table-global-options';
+import { AutoHeightDirective } from './directives/auto-height.directive';
 import { CalculateRange, ColumnsSchema } from './interfaces/table-builder.external';
 import { RecalculatedStatus, TableSimpleChanges, TemplateKeys } from './interfaces/table-builder.internal';
+import { getClientHeight } from './operators/get-client-height';
 import { ContextMenuService } from './services/context-menu/context-menu.service';
 import { DraggableService } from './services/draggable/draggable.service';
 import { FilterableService } from './services/filterable/filterable.service';
@@ -66,6 +70,8 @@ export class TableBuilderComponent<T>
 {
     @ViewChild('header', { static: false }) public headerRef!: ElementRef<HTMLDivElement>;
     @ViewChild('footer', { static: false }) public footerRef!: ElementRef<HTMLDivElement>;
+    @ViewChild(AutoHeightDirective, { static: false })
+    public readonly autoHeight!: AutoHeightDirective<T>;
     public dirty: boolean = true;
     public rendering: boolean = false;
     public isRendered: boolean = false;
@@ -113,6 +119,26 @@ export class TableBuilderComponent<T>
 
     public get sourceExists(): boolean {
         return this.sourceRef.length > 0;
+    }
+
+    public get rootHeight(): string {
+        const height: Nullable<string | number> = this.expandableTableExpanded
+            ? this.height
+            : getClientHeight(this.headerRef) + getClientHeight(this.footerRef);
+        if (checkValueIsFilled(height)) {
+            const heightAsNumber: number = Number(height);
+            return isNaN(heightAsNumber) ? String(height) : `${height}px`;
+        } else {
+            return '';
+        }
+    }
+
+    private get expandableTableExpanded(): boolean {
+        return (
+            isNil(this.headerTemplate) ||
+            !coerceBoolean(this.headerTemplate.expandablePanel) ||
+            coerceBoolean(this.headerTemplate.expanded)
+        );
     }
 
     private get viewIsDirty(): boolean {
@@ -202,6 +228,8 @@ export class TableBuilderComponent<T>
         if (this.sourceExists) {
             this.render();
         }
+
+        this.listenExpandChange();
     }
 
     public ngAfterViewInit(): void {
@@ -307,6 +335,11 @@ export class TableBuilderComponent<T>
         }
     }
 
+    public updateColumnsSchema(patch: PlainObjectOf<Partial<ColumnsSchema>>): void {
+        this.templateParser.updateColumnsSchema(patch);
+        this.changeSchema();
+    }
+
     public resetSchema(): void {
         this.columnListWidth = 0;
         this.schemaColumns = null;
@@ -340,6 +373,11 @@ export class TableBuilderComponent<T>
 
     public setSource(source: Nullable<T[]>): void {
         this.originalSource = this.source = this.selection.rows = source;
+    }
+
+    public updateTableHeight(): void {
+        this.autoHeight.calculateHeight();
+        detectChanges(this.cd);
     }
 
     protected calculateViewPortByRange({ start, end, bufferOffset, force }: CalculateRange): void {
@@ -749,7 +787,11 @@ export class TableBuilderComponent<T>
      * because users can draw the wrong keys in the template (ngx-column key=invalid)
      */
     private parseTemplateKeys(): TemplateKeys {
-        this.templateParser.keyMap = this.generateColumnsKeyMap(this.keys.length ? this.keys : this.getModelKeys());
+        const modelKeys: string[] = this.getModelKeys();
+        const keys: string[] = hasItems(this.keys)
+            ? this.keys.filter((key: string): boolean => modelKeys.includes(key))
+            : modelKeys;
+        this.templateParser.keyMap = this.generateColumnsKeyMap(keys);
 
         this.templateParser.allowedKeyMap = this.keys.length
             ? this.generateColumnsKeyMap(this.customModelColumnsKeys)
@@ -762,5 +804,11 @@ export class TableBuilderComponent<T>
             overridingRenderedKeys: this.templateParser.overrideTemplateKeys!,
             simpleRenderedKeys: this.templateParser.templateKeys!
         };
+    }
+
+    private listenExpandChange(): void {
+        this.headerTemplate?.expandedChange
+            .pipe(takeUntil(this.destroy$))
+            .subscribe((): void => this.updateTableHeight());
     }
 }